[12345] add permissions to students queries & mutation ...

+++ src/modules/graphql/resolvers/student.resolver.ts
@@ -1,7 +1,8 @@
 import { Resolver, Query, Mutation, ResolveProperty } from '@nestjs/graphql';
 import { UsersPersistenceService } from '../../persistence/users.persistence.service';
 import { ClassPersistenceService } from '../../persistence/class.persistence.service';
-import { UserRole } from '../../../models/user.db.model';
+import { UserDbModel, UserRole} from '../../../models/user.db.model';
+import { Asset, DBOperation, NO_PERMISSION, Permission, Permissions} from './permissionRules';
 
 @Resolver('Student')
 export class StudentResolver {
@@ -9,11 +10,31 @@ export class StudentResolver {
 
     @Query('students')
     async getStudents(_, {}, context) {
+        const permission = Permissions.checkAndGetBasePermission(context.user, DBOperation.READ, Asset.STUDENT);
+        if (permission === Permission.FORBID) {
+            throw new Error(NO_PERMISSION);
+        }
+        if (permission === Permission.OWN) {
+            // find students in teacher's class
+            const teacher: UserDbModel = await this.usersPersistence.getById(context.user.id);
+            return this.usersPersistence.getStudentsByClassId(teacher.class_id);
+        }
+
         return this.usersPersistence.getUsersByFilters({role: UserRole.STUDENT});
     }
 
     @Query('student')
     async getStudentById(_, args, context) {
+        const permission = Permissions.checkAndGetBasePermission(context.user, DBOperation.READ, Asset.STUDENT);
+        if (permission === Permission.FORBID) {
+            throw new Error(NO_PERMISSION);
+        }
+        if (permission === Permission.OWN) {
+            // find student in teacher's class
+            const teacher: UserDbModel = await this.usersPersistence.getById(context.user.id);
+            const [err, students] = await this.usersPersistence.getStudentsByClassId(teacher.class_id);
+            return students.find((obj) => obj._id === args);
+        }
         return this.usersPersistence.getUserByFilters({role: UserRole.STUDENT}, args.id);
     }
 
@@ -24,7 +45,10 @@ export class StudentResolver {
 
     @Mutation('createStudent')
     async createStudent(_, { student }, context) {
-        // TODO perform permissions rights
+        const permission = Permissions.checkAndGetBasePermission(context.user, DBOperation.CREATE, Asset.STUDENT);
+        if (permission === Permission.FORBID) {
+            throw new Error(NO_PERMISSION);
+        }
         // TODO: Handle errors!!!!
         const [, response] = await this.usersPersistence.createUser(student, UserRole.STUDENT);
         return response;
@@ -32,7 +56,19 @@ export class StudentResolver {
 
     @Mutation('updateStudent')
     async updateStudent(_, { id, student }, context) {
-        // TODO perform permissions rights
+        const permission = Permissions.checkAndGetBasePermission(context.user, DBOperation.UPDATE, Asset.STUDENT);
+        if (permission === Permission.FORBID) {
+            throw new Error(NO_PERMISSION);
+        }
+        if (permission === Permission.OWN) {
+            // find student in teacher's class
+            const teacher: UserDbModel = await this.usersPersistence.getById(context.user.id);
+            const [err, students] = await this.usersPersistence.getStudentsByClassId(teacher.class_id);
+            const studentInClass = students.find((obj) => obj._id === id);
+            if (!studentInClass) {
+                throw new Error(NO_PERMISSION);
+            }
+        }
         // TODO: Handle errors!!!!
         const [, response] = await this.usersPersistence.updateUser(id, student, UserRole.STUDENT);
         return response;
@@ -40,7 +76,19 @@ export class StudentResolver {
 
     @Mutation('deleteStudent')
     async deleteStudent(_, { id }, context) {
-        // TODO perform permissions rights
+        const permission = Permissions.checkAndGetBasePermission(context.user, DBOperation.DELETE, Asset.STUDENT);
+        if (permission === Permission.FORBID) {
+            throw new Error(NO_PERMISSION);
+        }
+        if (permission === Permission.OWN) {
+            // find student in teacher's class
+            const teacher: UserDbModel = await this.usersPersistence.getById(context.user.id);
+            const [err, students] = await this.usersPersistence.getStudentsByClassId(teacher.class_id);
+            const studentInClass = students.find((obj) => obj._id === id);
+            if (!studentInClass) {
+                throw new Error(NO_PERMISSION);
+            }
+        }
         // TODO: Handle errors!!!!
         const [, response] = await this.usersPersistence.deleteUser(id);
         return response;
